requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
marshmallow==3.20.1
psycopg2-binary==2.9.7
Flask-Migrate==4.0.4
Flask-Marshmallow==0.15.0
Flask-SocketIO==5.3.4
eventlet==0.33.3
gunicorn==20.1.0
